promptopt/requirements.txt
dspy-ai
pytest
